vc/amd/opensil/genoa_poc/memmap: use unsigned long for resource index
Use an unsigned long as resource index type instead of an int to match
the data type used for the index in the resource struct and the
functions to report the resources.

Reviewed-on: https://review.coreboot.org/c/coreboot/+/80265
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>

+++ src/vendorcode/amd/opensil/genoa_poc/memmap.c
@@ -84,7 +84,7 @@ static void print_memory_holes(void *unused)
 BOOT_STATE_INIT_ENTRY(BS_DEV_RESOURCES, BS_ON_ENTRY, print_memory_holes, NULL);
 
 // This assumes holes are allocated
-int add_opensil_memmap(struct device *dev, int idx)
+unsigned long add_opensil_memmap(struct device *dev, unsigned long idx)
 {
 	ram_from_to(dev, idx++, 0, 0xa0000);
 	mmio_from_to(dev, idx++, 0xa0000, 0xc0000); // legacy VGA
